fix ninja t2

+++ tasks/seq/Odintsov_M_CountingMismatchedCharactersStr/src/ops_seq.cpp
@@ -29,7 +29,7 @@ bool Odintsov_M_CountingMismatchedCharactersStr_seq::CountingCharacterSequential
 }
 bool Odintsov_M_CountingMismatchedCharactersStr_seq::CountingCharacterSequential::run() {
   internal_order_test();
-  for (int i = 0; i < strlen(input[0]); i++) {
+  for (size_t i = 0; i < strlen(input[0]); i++) {
     if (i < strlen(input[1])) {
       if (input[0][i] != input[1][i]) {
         ans += 2;
